pyproject.toml
[tool.poetry]
name = "stock-be"
version = "0.1.0"
description = ""
authors = ["Surakiet Piyasakulkiet <90365388+SurakietArt@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "stock_be"}]

[tool.isort]
skip = ["env", "stock_be/settings.py"]
line_length = 120
extend_skip_glob = ["scripts/unoconv"]
profile = "black"

[tool.black]
line-length = 88
exclude = '''
(
    ^manage\.py$
  | .+\\asgi\.py$
  | .+\\urls\.py$
  | .+\\wsgi\.py$
  | .+\\settings\.py$
)
'''

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
environ = "^1.0"
django-environ = "^0.11.2"
pre-commit = "^3.8.0"
isort = "^5.13.2"
psycopg2 = "^2.9.9"
django-safedelete = "^1.4.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
django-cors-headers = "^4.4.0"
django-dataclass-autoserialize = "^1.0.1"
google-api-python-client = "^2.147.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
